Switch to versioning via a file in the repo
We use pip in hashes mode in job-server which doesn't support installing
from a version control repo, even with setuptools-scm.

We want to make cutting a release something that is scripted in the
repo, such that if we do automate it, it's part of our just stack.

We've opted for calver her since we only care the version number on a
single axis (it gets bigger/smaller), and having some idea of when a
release was cut might be useful (but isn't a specific reason to have
picked it).

The version format we've picked here is:

    <year>.<month>.<day>.<hour><month><second>

Since this is a single `date` call it adds minimal infra to our stack.