Expose is_database_action logic as function
When running usable actions, which are *not* parsed by a Pipeline or
similar model atm, we need to detect whether a job is a db job (as
reusable actions are *not* allowed to be db jobs).

This change allows us to easily re-use the logic from outside of
a Pipeline model context.

Includes random black fixes for some reason :shrug: