Fix validation of `generate-dataset` privacy levels
As noted in a TODO comment in the code, the previous behaviour was just
flat out wrong. My suspicion is that this was a thinko/typo made in a
very early iteration of the validation code which then got ported to the
pipeline library out of a desire to exactly mirror the existing
behaviour.

In any case, what we want to enforce is that the outputs of any
`generate-dataset` actions are always labelled as highly sensitive, not
that they are labelled with only a single privacy level.