Refactor checking of `--output` argument
This is a bit of a fool's errand as it's unlikely we'll replicate every
single quirk of ehrQL's argument parsing (e.g. passing multiple
`--output` arguments) but we should at least cover the basics.

This check is intended as a time-saving early warning to researchers
that their code is broken rather than as any kind of permissions
enforcement. So it's not a disaster if doesn't handle all cases,
although if it throws a lot of false errors than that would be very
annoying.