Correctly handle multiple output files from ehrQL
I _really_ don't like the level of coupling between systems we've ended
up with here. But now doesn't feel like quite the right time to have the
wider discussion about whether we want these kind of validation checks
at all and, if so, whether the pipeline library is the place to have
them. So for now I'm just trying to preserve existing functionality in
the least-worst way I can.